feat: combine controllers to one deployment
