+++ manifests/base/controller-manager/controller-manager-deployment.yaml
@@ -1,32 +1,34 @@
 apiVersion: apps/v1
 kind: Deployment
 metadata:
-  name: eventbus-controller
+  name: controller-manager
 spec:
   replicas: 1
   selector:
     matchLabels:
-      app: eventbus-controller
+      app: controller-manager
   template:
     metadata:
       labels:
-        app: eventbus-controller
+        app: controller-manager
     spec:
       serviceAccountName: argo-events-sa
       securityContext:
         runAsNonRoot: true
         runAsUser: 9731
       containers:
-        - name: eventbus-controller
+        - name: main
           image: quay.io/argoproj/argo-events:latest
           imagePullPolicy: Always
           args:
-            - eventbus-controller
+            - controller
           env:
             - name: NAMESPACE
               valueFrom:
                 fieldRef:
                   fieldPath: metadata.namespace
+            - name: ARGO_EVENTS_IMAGE
+              value: quay.io/argoproj/argo-events:latest
           volumeMounts:
             - mountPath: /etc/argo-events
               name: controller-config-volume
