+++ manifests/base/controller-manager/kustomization.yaml
@@ -2,4 +2,5 @@ apiVersion: kustomize.config.k8s.io/v1beta1
 kind: Kustomization
 
 resources:
-  - eventsource-controller-deployment.yaml
+  - controller-config.yaml
+  - controller-manager-deployment.yaml
