--- manifests/base/eventbus-controller/kustomization.yaml
@@ -1,6 +1,0 @@
-apiVersion: kustomize.config.k8s.io/v1beta1
-kind: Kustomization
-
-resources:
-  - controller-config.yaml
-  - eventbus-controller-deployment.yaml
