--- manifests/base/sensor-controller/sensor-controller-deployment.yaml
@@ -1,43 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: sensor-controller
-spec:
-  replicas: 1
-  selector:
-    matchLabels:
-      app: sensor-controller
-  template:
-    metadata:
-      labels:
-        app: sensor-controller
-    spec:
-      serviceAccountName: argo-events-sa
-      securityContext:
-        runAsNonRoot: true
-        runAsUser: 9731
-      containers:
-        - name: sensor-controller
-          image: quay.io/argoproj/argo-events:latest
-          imagePullPolicy: Always
-          args:
-            - sensor-controller
-          env:
-            - name: NAMESPACE
-              valueFrom:
-                fieldRef:
-                  fieldPath: metadata.namespace
-            - name: SENSOR_IMAGE
-              value: quay.io/argoproj/argo-events:latest
-          livenessProbe:
-            httpGet:
-              path: /healthz
-              port: 8081
-            initialDelaySeconds: 3
-            periodSeconds: 3
-          readinessProbe:
-            httpGet:
-              path: /readyz
-              port: 8081
-            initialDelaySeconds: 3
-            periodSeconds: 3
