+++ manifests/install.yaml
@@ -314,6 +314,58 @@ metadata:
 apiVersion: apps/v1
 kind: Deployment
 metadata:
+  name: controller-manager
+  namespace: argo-events
+spec:
+  replicas: 1
+  selector:
+    matchLabels:
+      app: controller-manager
+  template:
+    metadata:
+      labels:
+        app: controller-manager
+    spec:
+      containers:
+      - args:
+        - controller
+        env:
+        - name: NAMESPACE
+          valueFrom:
+            fieldRef:
+              fieldPath: metadata.namespace
+        - name: ARGO_EVENTS_IMAGE
+          value: quay.io/argoproj/argo-events:latest
+        image: quay.io/argoproj/argo-events:latest
+        imagePullPolicy: Always
+        livenessProbe:
+          httpGet:
+            path: /healthz
+            port: 8081
+          initialDelaySeconds: 3
+          periodSeconds: 3
+        name: main
+        readinessProbe:
+          httpGet:
+            path: /readyz
+            port: 8081
+          initialDelaySeconds: 3
+          periodSeconds: 3
+        volumeMounts:
+        - mountPath: /etc/argo-events
+          name: controller-config-volume
+      securityContext:
+        runAsNonRoot: true
+        runAsUser: 9731
+      serviceAccountName: argo-events-sa
+      volumes:
+      - configMap:
+          name: argo-events-controller-config
+        name: controller-config-volume
+---
+apiVersion: apps/v1
+kind: Deployment
+metadata:
   name: eventbus-controller
   namespace: argo-events
 spec:
@@ -349,17 +401,10 @@ spec:
             port: 8081
           initialDelaySeconds: 3
           periodSeconds: 3
-        volumeMounts:
-        - mountPath: /etc/argo-events
-          name: controller-config-volume
       securityContext:
         runAsNonRoot: true
         runAsUser: 9731
       serviceAccountName: argo-events-sa
-      volumes:
-      - configMap:
-          name: argo-events-controller-config
-        name: controller-config-volume
 ---
 apiVersion: apps/v1
 kind: Deployment
