+++ test/manifests/kustomization.yaml
@@ -12,21 +12,7 @@ patches:
         value: IfNotPresent
     target:
       kind: Deployment
-      name: eventbus-controller
-  - patch: |-
-      - op: replace
-        path: /spec/template/spec/containers/0/imagePullPolicy
-        value: IfNotPresent
-    target:
-      kind: Deployment
-      name: eventsource-controller
-  - patch: |-
-      - op: replace
-        path: /spec/template/spec/containers/0/imagePullPolicy
-        value: IfNotPresent
-    target:
-      kind: Deployment
-      name: sensor-controller
+      name: controller-manager
   - patch: |-
       - op: replace
         path: /spec/template/spec/containers/0/imagePullPolicy
@@ -42,16 +28,7 @@ patches:
           value: IfNotPresent
     target:
       kind: Deployment
-      name: eventsource-controller
-  - patch: |-
-      - op: add
-        path: /spec/template/spec/containers/0/env/-
-        value:
-          name: IMAGE_PULL_POLICY
-          value: IfNotPresent
-    target:
-      kind: Deployment
-      name: sensor-controller
+      name: controller-manager
 
 namespace: argo-events
